add fit job-submisstion

+++ preparehistojobs.py
@@ -7,25 +7,27 @@ import yaml
 def prepare_job_script():
 		plotDirectory = jobsDirectory +"/"+year
 		if not os.path.exists(plotDirectory):
-			os.makedirs(plotDirectory)
+      os.makedirs(plotDirectory)
+    if not os.path.exists(jobsDirectory +"/jobs"):
+      os.makedirs(jobsDirectory +"/jobs")
 		os.system("cp "+baseDir+"/bin/example_ntuple " + jobsDirectory)
 		os.system("cp "+baseDir+"/env.sh " + jobsDirectory)
 		subDirectory = jobsDirectory +"/"+year+"/histograms"
 		if not os.path.exists(subDirectory):
 			os.makedirs(subDirectory)
 		for fileName in samples:
-			scriptFile = open(jobsDirectory+'/'+'mkhis_'+year+fileName+'.sh','w')
+			scriptFile = open(jobsDirectory+'/jobs/'+'mkhis_'+year+fileName+'.sh','w')
 			scriptLines = ''
 			scriptLines += ('export INITDIR='+jobsDirectory+'\n')
 			scriptLines += ('cd $INITDIR\n')
 			scriptLines += '. ./env.sh ;\n'
 			scriptLines += ("date;\n")
-			scriptLines += ("./example_ntuple input="+baseDir+"/samples/"+year+fileName+".root"+" config="+baseDir+"/config/"+aFile+" output="+subDirectory+fileName+".root;\n")
+			scriptLines += ("./example_ntuple input="+baseDir+"/samples/"+year+fileName+".root"+" config="+baseDir+"/config/"+aFile+" output="+subDirectory+"/"+fileName+".root;\n")
 			#scriptLines += ("./fit.py -i "+subDirectory+fileName+".root" +" -s "+fileName+" -y "+baseDir+"/config/"+aFile+" -f "+aModel+" -o "+subDirectory+fileName+"_eff.root;\n")
 			scriptFile.write(scriptLines)
 			scriptFile.close()
-			jobsFiles = open(jobsDirectory+"/sendJobs.cmd","a")
-			jobsFiles.write("qsub  "+jobsDirectory+'/'+'mkhis_'+year+fileName+'.sh\n')
+			jobsFiles = open(jobsDirectory+"/jobs/sendJobs.cmd","a")
+			jobsFiles.write("qsub  "+jobsDirectory+'/jobs/mkhis_'+year+fileName+'.sh\n')
 			jobsFiles.close()
 
 def main():
@@ -35,13 +37,14 @@ def main():
     jobsDirectory = baseDir+"/outputs/"+label 
     if not os.path.exists(jobsDirectory):
       os.makedirs(jobsDirectory)
+    if not os.path.exists(jobsDirectory+"/jobs"):
+      os.makedirs(jobsDirectory+"/jobs")
     prepare_job_script()
-	
 
 if __name__ == '__main__':
   files = os.listdir(os.getcwd()+"/config")
   for aFile in files:
-    f = open(aFile)
+    f = open("config/"+aFile)
     config = yaml.load(f)
     yaml.dump(config)
 
@@ -50,5 +53,5 @@ if __name__ == '__main__':
     samples = config['sample']
     print(samples)
     pdfs = config['pdfs']
-    main()
-
+    main()  
+  os.system("big-submission "+jobsDirectory+"/jobs/sendJobs.cmd")
